Fix rounding error in ParStripProjector2D::getProjectionWeightsCount
In the worst case this would lead to (nearly) empty storage for getMatrix(),
resulting in (near) explicit projection matrices. (These are only used for
exporting explicit sparse projection matrices to matlab/python; not for
FP/BP/reconstruction.)

This is a quick fix; ideally the affected code would use dynamic storage.

+++ src/ParallelBeamStripKernelProjector2D.cpp
@@ -141,8 +141,8 @@ bool CParallelBeamStripKernelProjector2D::initialize(CParallelProjectionGeometry
 int CParallelBeamStripKernelProjector2D::getProjectionWeightsCount(int _iProjectionIndex)
 {
 	int maxDim = max(m_pVolumeGeometry->getGridRowCount(), m_pVolumeGeometry->getGridColCount());
-	int scale = m_pProjectionGeometry->getDetectorWidth() / min(m_pVolumeGeometry->getPixelLengthX(), m_pVolumeGeometry->getPixelLengthY());
-	return maxDim * scale * 10 + 1;
+	double scale = m_pProjectionGeometry->getDetectorWidth() / min(m_pVolumeGeometry->getPixelLengthX(), m_pVolumeGeometry->getPixelLengthY());
+	return int(maxDim * scale * 10) + 1;
 }
 
 //----------------------------------------------------------------------------------------
